tile_dataset/requirements.txt
requests==2.31.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
pillow==10.1.0
piexif==1.1.3
